Feature/9 upload log file

+++ backend/src/main/java/com/humuson/backend/domain/log/model/response/LogAnalysisResponse.java
@@ -4,9 +4,9 @@ import lombok.AccessLevel;
 import lombok.Builder;
 
 @Builder(access = AccessLevel.PRIVATE)
-public record LogAnalysisResponse(int totalLogsCount, long errorLogsCount) {
+public record LogAnalysisResponse(long totalLogsCount, long errorLogsCount) {
 
-    public static LogAnalysisResponse of(int totalLogsCount, long errorLogsCount) {
+    public static LogAnalysisResponse of(long totalLogsCount, long errorLogsCount) {
         return LogAnalysisResponse.builder()
                 .totalLogsCount(totalLogsCount)
                 .errorLogsCount(errorLogsCount)
